refactor(directory): change directory structure
