+++ src/app.ts
@@ -4,8 +4,8 @@ import * as logger from 'morgan';
 import * as bodyParser from 'body-parser';
 import * as mongoose from 'mongoose';
 
-import { userRoutes } from './api/routes/user-routes';
-import { carInfoRoutes } from './api/routes/car-info-routes';
+import { userRoutes } from './api/v1/user/user-routes';
+import { carInfoRoutes } from './api/v1/car-info/car-info-routes';
 
 export const app: express.Application = express();
 
@@ -48,7 +48,6 @@ this.app.use((req, res, next) => {
 // Routes to handle requests
 this.app.use('/users', userRoutes);
 this.app.use('/cars', carInfoRoutes)
-this.app.use('/carsOwned', )
 
 
 this.app.use((req, res, next) => {
